Change in Date type from long to double. New method added 'StandardBlack::evaluate'

+++ include/Date.h
@@ -5,7 +5,7 @@ namespace ql
 {
 	// Type for dates, expressed in Excel convention. 
 	// For instance 42003 is 30 December 2004
-	typedef long Date;
+	typedef double Date;
 
 
 }
